+++ src/main/scala/examples/Risc.scala
@@ -16,8 +16,8 @@ class Risc extends Module {
   val file = Mem(256, UInt(32.W))
   val code = Mem(256, UInt(32.W))
   val pc   = Reg(init=0.U(8.W))
-  
-  val add_op :: imm_op :: Nil = Enum(UInt(), 2)
+
+  val add_op :: imm_op :: Nil = Enum(2)
 
   val inst = code(pc)
   val op   = inst(31,24)
